Made it so a user can remove a person from the list - need to get it to auto do it when the button is pressed
