+++ wordBombHelper/wordBombHelper/App.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using wordBombHelper.Views;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
@@ -13,7 +14,11 @@ namespace wordBombHelper
         {
             InitializeComponent();
 
-            MainPage = new MainPage();
+            MainPage = new NavigationPage(new PlayerPage())
+            {
+                BarTextColor = Color.White,
+                BarBackgroundColor = Color.Aquamarine
+            };
         }
 
         protected override void OnStart()
