+++ wordBombHelper/wordBombHelper/MainPage.xaml
@@ -6,44 +6,6 @@
     <StackLayout Margin="20,35,20,20">
 
         <Label Text="Enter up to four players below:"/>
-        <Entry x:Name="playerOneEntry"
-               Placeholder="Player one"
-                />
-        <Button Text="Add player"
-                Clicked="OnPlayerOneClicked"/>
-        <Entry x:Name="playerTwoEntry"
-                Placeholder="Player two"/>
-        <Button Text="Add player"
-                Clicked="OnPlayerTwoClicked"/>
-        <Entry x:Name="playerThreeEntry"
-                Placeholder="Player three"/>
-        <Button Text="Add player"
-                Clicked="OnPlayerThreeClicked"/>
-        <Entry x:Name="playerFourEntry"
-                Placeholder="Player four"/>
-        <Button Text="Add player"
-                Clicked="OnPlayerFourClicked"/>
-
-        <ListView x:Name="allPlayers">
-            <ListView.ItemTemplate>
-                <DataTemplate>
-                    <ViewCell>
-                        <StackLayout Margin="20,0,0,0"
-                                     Orientation="Horizontal"
-                                     HorizontalOptions="FillAndExpand">
-                            <Label Text="{Binding PlayerName}"
-                                   VerticalTextAlignment="Center"
-                                   HorizontalOptions="StartAndExpand"/>
-                            <Label Text="{Binding Colour}"
-                                   VerticalTextAlignment="Center"
-                                   TextColor="{Binding Colour}"/>
-                            
-                        </StackLayout>
-                    </ViewCell>
-                </DataTemplate>
-            </ListView.ItemTemplate>
-        </ListView>
 
     </StackLayout>
-
 </ContentPage>
